+++ go.mod
@@ -107,8 +107,10 @@ require (
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 	k8s.io/apiextensions-apiserver v0.34.0-beta.0 // indirect
 	k8s.io/cloud-provider v0.34.0-beta.0 // indirect
+	k8s.io/code-generator v0.34.0-beta.0 // indirect
 	k8s.io/component-base v0.34.0-beta.0 // indirect
 	k8s.io/csi-translation-lib v0.34.0-beta.0 // indirect
+	k8s.io/gengo/v2 v2.0.0-20250604051438-85fd79dbfd9f // indirect
 	k8s.io/kube-openapi v0.0.0-20250710124328-f3f2b991d03b // indirect
 	sigs.k8s.io/json v0.0.0-20241014173422-cfa47c3a1cc8 // indirect
 	sigs.k8s.io/randfill v1.0.0 // indirect
